[api] Include CommentEvent in the Event module
To allow that the classes which includes it can find it, as now we do
things in the Rails way.
